Split CarryUnit into PickupUnit and DeliverUnit
